Modifying the context such  that I can get it in a state to push serialization

+++ source/RTSCamera.cpp
@@ -132,7 +132,7 @@ void RTSCamera::OnRegisterSceneNode()
 
    f32 dp = tgtv.dotProduct(up); 
 
-   if ( core::equals ( fabs ( dp ), 1.f ) ) 
+   if ( core::equals ( (f32)fabs ( dp ), 1.f ) ) 
    { 
       up.X += 0.5f; 
    }
